[Event Hubs Client] Event Processor<T> Main Loop
The focus of these changes is to implement the main loop of the processor that
executes in the background when the processor is running, and is responsible for
coordinating with the load balancer for partition ownership changes and acting
as a sentinal for the active partition processors.
